refactor folder hierarchy. add dark/light mode support. refactor button component
