+++ lib/juno-ui-components/.storybook/main.js
@@ -1,4 +1,4 @@
 module.exports = {
-  stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  addons: ["@storybook/addon-links", "@storybook/addon-essentials"],
+  stories: ["../src/**/*.stories.@(js|jsx|ts|tsx|mdx)"],
+  addons: ["@storybook/addon-links", "@storybook/addon-essentials", "storybook-dark-mode/register"],
 }
