+++ lib/juno-ui-components/.storybook/preview.js
@@ -1,16 +1,25 @@
 import React from "react"
-import { GlobalStyles } from "twin.macro"
+import GlobalStyles from "/src/lib/styling/GlobalStyles"
+import { useDarkMode } from 'storybook-dark-mode'
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
+  backgrounds: { disable: true },
+  darkMode: {
+    stylePreview: true
+  }
 }
 
+
 export const decorators = [
-  (Story) => (
-    <div>
-      {/* */}
-      <GlobalStyles />
-      <Story />
-    </div>
-  ),
+  (Story) => {
+    const mode = useDarkMode() ? 'theme-dark' : 'theme-light'
+    
+    return (
+      <div className={mode}>
+        <GlobalStyles />
+        <Story />
+      </div>
+    )
+  },
 ]
