+++ lib/juno-ui-components/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
+    "polished": "^4.1.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1"
   },
@@ -60,10 +61,13 @@
     "@babel/preset-react": "^7.12.13",
     "@emotion/babel-plugin-jsx-pragmatic": "^0.1.5",
     "@emotion/cache": "^11.1.3",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.1.5",
     "@rollup/plugin-babel": "^5.3.0",
     "@storybook/addon-actions": "^6.1.21",
     "@storybook/addon-essentials": "^6.1.21",
     "@storybook/addon-links": "^6.1.21",
+    "@storybook/addons": "^6.2.8",
     "@storybook/react": "^6.1.21",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
@@ -80,6 +84,7 @@
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-includepaths": "^0.2.4",
     "rollup-plugin-root-import": "^1.0.0",
+    "storybook-dark-mode": "^1.0.7",
     "tailwindcss": "^2.1.1",
     "twin.macro": "^2.3.2"
   }
