+++ lib/juno-ui-components/src/components/Form/FloatingLabelInput.component.js
