+++ lib/juno-ui-components/src/components/Form/FloatingLabelSelect.component.js
