+++ lib/juno-ui-components/src/components/Form/Form.FloatingLabelInput.stories.js
@@ -24,10 +24,10 @@
 
 import React from "react"
 
-import { FloatingLabelInput } from "../components/Form/index.js"
+import { FloatingLabelInput } from "./index.js"
 
 export default {
-  title: "Example/Form/FloatingLabelInput",
+  title: "Design System/Components/Form/FloatingLabelInput",
   component: FloatingLabelInput,
   argTypes: {
     name: { control: "text" },
