+++ lib/juno-ui-components/src/components/Form/Form.FloatingLabelSelect.stories.js
@@ -24,10 +24,10 @@
 
 import React from "react"
 
-import { FloatingLabelSelect } from "../components/Form/index.js"
+import { FloatingLabelSelect } from "./index.js"
 
 export default {
-  title: "Example/Form/FloatingLabelSelect",
+  title: "Design System/Components/Form/FloatingLabelSelect",
   component: FloatingLabelSelect,
   argTypes: {
     name: { control: "text" },
