+++ lib/juno-ui-components/src/components/Modal/Modal.stories.js
@@ -1,11 +1,11 @@
 import React from "react"
 
-import { Modal } from "../components/Modal/index.js"
-import { Button } from "../components/Button/index.js"
-import { FloatingLabelInput } from "../components/Form/index.js"
+import Modal from "./index.js"
+import Button from "../../primitives/Button/index.js"
+import { FloatingLabelInput } from "../Form/index.js"
 
 export default {
-  title: "Example/Modal",
+  title: "Design System/Components/Modal",
   component: Modal,
   argTypes: {
     title: { control: "text" },
