+++ lib/juno-ui-components/src/primitives/Button/index.js
@@ -1,0 +1,2 @@
+import Button from './Button.component'
+export default Button
