+++ lib/juno-ui-components/src/primitives/Spinner/Spinner.component.js
@@ -21,7 +21,7 @@ const defaultColor = tw`
   text-white
 `
 
-export const Spinner = ({ color }) => {
+const Spinner = ({ color }) => {
   const mode = () => {
     switch (color) {
       case "primary":
@@ -60,3 +60,5 @@ export const Spinner = ({ color }) => {
     </svg>
   )
 }
+
+export default Spinner
