+++ lib/juno-ui-components/src/primitives/Spinner/Spinner.stories.js
@@ -1,9 +1,9 @@
 import React from "react"
 
-import { Spinner } from "../components/Spinner/index.js"
+import { Spinner } from "./index.js"
 
 export default {
-  title: "Example/Spinner",
+  title: "Design System/Primitives/Spinner",
   component: Spinner,
   argTypes: {
     // backgroundColor: { control: "color" },
