+++ lib/juno-ui-components/src/primitives/Spinner/index.js
@@ -1,0 +1,2 @@
+import Spinner from "./Spinner.component"
+export default Spinner
