+++ package.json
@@ -18,6 +18,7 @@
     }
   },
   "dependencies": {
+    "@ibm/plex": "^5.1.3",
     "react-static": "^7.5.3"
   }
 }
